+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 ### Added
 * Added camera module
 * Added deltatime
+* Added lots more functionality to the engine
+* Added texture class
+* Added cube instancing
+
+### Changed
+* Moved shader class to graphics.cpp
+* No need for a cleanup function (yet) all class object's deconstructors will do the cleanup
 
 -------------------------------------------------------------------------------------------------------------